Decompose unit tests into unit, conv, and sim
Remove OrdinaryDiffEq from test extras
