+++ test/Operators/hybrid/utils_3d.jl
@@ -1,0 +1,65 @@
+using Test
+using ClimaComms
+ClimaComms.@import_required_backends
+using StaticArrays, IntervalSets, LinearAlgebra
+
+import ClimaCore:
+    ClimaCore,
+    slab,
+    Domains,
+    Meshes,
+    Geometry,
+    Topologies,
+    Grids,
+    Spaces,
+    Quadratures,
+    Fields,
+    Operators
+import ClimaCore.Geometry: WVector
+
+import ClimaCore.Utilities: half
+import ClimaCore.DataLayouts: level
+
+function hvspace_3D(
+    xlim = (-π, π),
+    ylim = (-π, π),
+    zlim = (0, 4π),
+    xelem = 4,
+    yelem = 4,
+    zelem = 16,
+    npoly = 7;
+    device = ClimaComms.device(),
+)
+    FT = Float64
+    vertdomain = Domains.IntervalDomain(
+        Geometry.ZPoint{FT}(zlim[1]),
+        Geometry.ZPoint{FT}(zlim[2]);
+        boundary_names = (:bottom, :top),
+    )
+    vertmesh = Meshes.IntervalMesh(vertdomain, nelems = zelem)
+    verttopology = Topologies.IntervalTopology(
+        ClimaComms.SingletonCommsContext(device),
+        vertmesh,
+    )
+    vert_center_space = Spaces.CenterFiniteDifferenceSpace(verttopology)
+
+    horzdomain = Domains.RectangleDomain(
+        Geometry.XPoint{FT}(xlim[1]) .. Geometry.XPoint{FT}(xlim[2]),
+        Geometry.YPoint{FT}(ylim[1]) .. Geometry.YPoint{FT}(ylim[2]),
+        x1periodic = true,
+        x2periodic = true,
+    )
+    horzmesh = Meshes.RectilinearMesh(horzdomain, xelem, yelem)
+    horztopology = Topologies.Topology2D(
+        ClimaComms.SingletonCommsContext(device),
+        horzmesh,
+    )
+
+    quad = Quadratures.GLL{npoly + 1}()
+    horzspace = Spaces.SpectralElementSpace2D(horztopology, quad)
+
+    hv_center_space =
+        Spaces.ExtrudedFiniteDifferenceSpace(horzspace, vert_center_space)
+    hv_face_space = Spaces.FaceExtrudedFiniteDifferenceSpace(hv_center_space)
+    return (hv_center_space, hv_face_space)
+end
